# services/hotel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hotel microservice.",
        "title": "Booking project",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/hotels/all": {
            "get": {
                "description": "Get list of hotels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get list of hotels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "City location",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Stars",
                        "name": "stars",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Hotel"
                            }
                        }
                    }
                }
            }
        },
        "/hotels/create": {
            "post": {
                "description": "Add new hotel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Add new hotel",
                "parameters": [
                    {
                        "description": "Hotel data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateHotelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Hotel"
                        }
                    }
                }
            }
        },
        "/hotels/update/{id}": {
            "put": {
                "description": "Update hotel information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Update hotel information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hotel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hotel data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateHotelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/hotels/{id}": {
            "get": {
                "description": "Get hotel information by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get hotel by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Hotel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Hotel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CreateHotelRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "name",
                "stars"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 100
                },
                "city": {
                    "type": "string",
                    "maxLength": 32
                },
                "name": {
                    "type": "string",
                    "maxLength": 64
                },
                "stars": {
                    "type": "integer",
                    "maximum": 7,
                    "minimum": 0
                }
            }
        },
        "domain.Hotel": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stars": {
                    "type": "integer"
                }
            }
        }
    }
}